Rewrite for new behaviour and AngularJS integration
This is motivated by my work in Viroverse to provide a unified sequence
and chromats upload field for each positive PCR product instead of
separating the sequences from the chromats.

It changes the behaviour of file selection and file drops to accumulate
instead of replace, makes it possible to clear the upload list, adds a
"click to add files" behaviour, and provides easy integration with
AngularJS.  The overall goal of a loosely coupled component using
vanilla elements and JS with a lot of flexibility in UI is still
preserved.